pepperLife/requirements.txt
openai>=1.30.0
pypng>=0.20220715.0